artifact/scripts/api-tools/tools_out/python/requirements.txt
# Core dependencies
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
typing-extensions>=4.8.0

# API and Schema validation
openapi-spec-validator>=0.7.1
jsonschema>=4.20.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Utility
python-dateutil>=2.8.2
tqdm>=4.66.1

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# Development
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.1 